Jupyter notebook notes added
I learned basic jupyter notebook use w/ a markdown refresher and copied some of my notes to from my notion site to get comfortable with it.
I figured eventually I'll need to know how to use jupyter notebook. I've watched a basic video of what it can do and have seen it mentioned literally everywhere so I might as well get comfortable using it now in a way that's simple so that when things get complicated I won't have to learn both the tool and it's functions